Update styles.css/Circling An Image and Device Adaptation
Well i managed to circle the image and also center it. it gave me quite a few problems which led me to re adjust the nav links properties because the image affected it, but I'm still struggling to position the outer circle well I'll go again tommorow by the grace of God. Chatgpt showed me a query for device adaptability because my mobile view was looking messy.